refactor(agent-run): rename helper and extract run store
